requirements.txt

fastapi
uvicorn
pandas
pandas-ta
requests
python-dotenv
